Move actions unit_tests to the correct folder
The unit_tests for system_checks and pre_ponr_changes were in the
top-level `actions` folder, which is not where they belong anymore.
