+++ convert2rhel/unit_tests/actions/system_checks/efi_test.py
