+++ convert2rhel/unit_tests/actions/system_checks/is_loaded_kernel_latest_test.py
