+++ convert2rhel/unit_tests/actions/system_checks/rhel_compatible_kernel_test.py
